Completing renaming from eclipsing -> periodic
