+++ astroplan/__init__.py
@@ -27,6 +27,6 @@ if not _ASTROPY_SETUP_:
     from .moon import *
     from .constraints import *
     from .scheduling import *
-    from .eclipsing import *
+    from .periodic import *
 
     get_IERS_A_or_workaround()
